Optimize tool hallucination handling (#3773)
