+++ langchain4j/src/test/java/dev/langchain4j/service/AiServicesUserMessageConfigTest.java
@@ -8,21 +8,13 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.verify;
 
 import dev.langchain4j.invocation.InvocationParameters;
-import dev.langchain4j.agent.tool.Tool;
-import dev.langchain4j.agent.tool.ToolExecutionRequest;
 import dev.langchain4j.data.image.Image;
-import dev.langchain4j.data.message.AiMessage;
-import dev.langchain4j.data.message.ChatMessage;
 import dev.langchain4j.data.message.Content;
 import dev.langchain4j.data.message.ImageContent;
 import dev.langchain4j.data.message.TextContent;
-import dev.langchain4j.data.message.ToolExecutionResultMessage;
-import dev.langchain4j.memory.ChatMemory;
-import dev.langchain4j.memory.chat.MessageWindowChatMemory;
 import dev.langchain4j.model.chat.ChatModel;
 import dev.langchain4j.model.chat.mock.ChatModelMock;
 import dev.langchain4j.model.chat.request.ChatRequest;
-import dev.langchain4j.service.tool.HallucinatedToolNameStrategy;
 import java.util.List;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
@@ -522,105 +514,5 @@ class AiServicesUserMessageConfigTest {
         assertThatThrownBy(() -> aiService.illegalChat8("Hello", invocationParameters, invocationParameters))
                 .isExactlyInstanceOf(IllegalConfigurationException.class)
                 .hasMessage("There can be at most one parameter of type dev.langchain4j.invocation.InvocationParameters");
-    }
-
-    interface AssistantHallucinatedTool {
-        Result<AiMessage> chat(String userMessage);
-    }
-
-    static class HelloWorld {
-
-        @Tool("Say hello")
-        String hello(String name) {
-            return "Hello " + name + "!";
-        }
-    }
-
-    @Test
-    void should_fail_on_hallucinated_tool_execution() {
-
-        ChatModel chatModel = new ChatModelMock(ignore -> AiMessage.from(
-                ToolExecutionRequest.builder().id("id").name("unknown").build()));
-
-        ChatMemory chatMemory = MessageWindowChatMemory.withMaxMessages(10);
-
-        AssistantHallucinatedTool assistant = AiServices.builder(AssistantHallucinatedTool.class)
-                .chatModel(chatModel)
-                .chatMemory(chatMemory)
-                .tools(new HelloWorld())
-                .hallucinatedToolNameStrategy(HallucinatedToolNameStrategy.THROW_EXCEPTION)
-                .build();
-
-        assertThatThrownBy(() -> assistant.chat("hi"))
-                .isExactlyInstanceOf(RuntimeException.class)
-                .hasMessageContaining("unknown");
-
-        validateChatMemory(chatMemory);
-    }
-
-    @Test
-    void should_retry_on_hallucinated_tool_execution() {
-
-        ChatModel chatModel = new ChatModelMock(chatRequest -> {
-            List<ToolExecutionResultMessage> toolResults = chatRequest.messages().stream()
-                    .filter(ToolExecutionResultMessage.class::isInstance)
-                    .map(ToolExecutionResultMessage.class::cast)
-                    .toList();
-            if (toolResults.isEmpty()) {
-                return AiMessage.from(
-                        ToolExecutionRequest.builder().id("id").name("unknown").build());
-            }
-            ToolExecutionResultMessage lastToolResult = toolResults.get(toolResults.size() - 1);
-            String text = lastToolResult.text();
-            if (text.contains("Error")) {
-                // The LLM is supposed to understand the error and retry with the correct tool name
-                return AiMessage.from(ToolExecutionRequest.builder()
-                        .id("id")
-                        .name("hello")
-                        .arguments("{\"arg0\": \"Mario\"}")
-                        .build());
-            }
-            return AiMessage.from(text);
-        });
-
-        ChatMemory chatMemory = MessageWindowChatMemory.withMaxMessages(10);
-
-        AssistantHallucinatedTool assistant = AiServices.builder(AssistantHallucinatedTool.class)
-                .chatModel(chatModel)
-                .chatMemory(chatMemory)
-                .tools(new HelloWorld())
-                .hallucinatedToolNameStrategy(toolExecutionRequest -> ToolExecutionResultMessage.from(
-                        toolExecutionRequest, "Error: there is no tool called " + toolExecutionRequest.name()))
-                .build();
-
-        Result<AiMessage> result = assistant.chat("hi");
-        assertThat(result.content().text()).isEqualTo("Hello Mario!");
-
-        validateChatMemory(chatMemory);
-    }
-
-    private static void validateChatMemory(ChatMemory chatMemory) {
-        List<ChatMessage> messages = chatMemory.messages();
-        Class<?> expectedMessageType = dev.langchain4j.data.message.UserMessage.class;
-        for (ChatMessage message : messages) {
-            assertThat(message).isInstanceOf(expectedMessageType);
-            expectedMessageType = nextExpectedMessageType(message);
-        }
-    }
-
-    private static Class<?> nextExpectedMessageType(ChatMessage message) {
-        if (message instanceof dev.langchain4j.data.message.UserMessage) {
-            return AiMessage.class;
-        } else if (message instanceof AiMessage aiMessage) {
-            if (aiMessage.hasToolExecutionRequests()) {
-                return ToolExecutionResultMessage.class;
-            } else {
-                return dev.langchain4j.data.message.UserMessage.class;
-            }
-        } else if (message instanceof ToolExecutionResultMessage) {
-            return AiMessage.class;
-        }
-        throw new UnsupportedOperationException(
-                "Unsupported message type: " + message.getClass().getName());
     }
 }
